backend/pyproject.toml
[project]
name = "authority_tracker"
version = "0.1.0"
description = "A FastAPI project"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "your@email.com" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "orjson>=3.10.0",
    "logfire[fastapi]>=2.0.0",
    "sqlalchemy[aiosqlite]>=2.0.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "fastapi-pagination>=0.12.31",
    "taskiq>=0.11.0",
    "taskiq-redis>=1.0.0",
    "taskiq-dependencies>=0.1.0",
    "click>=8.1.0",
    "tabulate>=0.9.0",
    "aiosqlite>=0.22.1",
    "python-socketio>=5.16.1",
    "google-genai>=1.64.0",
    "google-cloud-texttospeech>=2.34.0",
    "nest-asyncio>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "anyio[trio]>=4.0.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
authority_tracker = "cli.commands:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "cli"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (needed for FastAPI Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["app", "cli"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.fastapi-fullstack]
generator_version = "0.1.18"
generated_at = "2026-02-21T14:45:21.907826+00:00"
